+++ demo/app.component.ts
@@ -20,17 +20,17 @@ export class AppComponent {
   ];
 
   nestedLists = [
-    { label: 'Item 1', children: [] },
+    { label: 'Item 1', children: [] as any[] },
     { label: 'Item 2', children: [
-        { label: 'Item 2a', children: [] },
-        { label: 'Item 2b', children: [] },
-        { label: 'Item 2c', children: [] },
+        { label: 'Item 2a', children: [] as any[] },
+        { label: 'Item 2b', children: [] as any[] },
+        { label: 'Item 2c', children: [] as any[] },
       ]
     },
     { label: 'Item 3', children: [
-        { label: 'Item 3a', children: [] },
-        { label: 'Item 3b', children: [] },
-        { label: 'Item 3c', children: [] },
+        { label: 'Item 3a', children: [] as any[] },
+        { label: 'Item 3b', children: [] as any[] },
+        { label: 'Item 3c', children: [] as any[] },
       ]
     },
   ];
@@ -40,12 +40,12 @@ export class AppComponent {
     { label: 'Item 2' },
     { label: 'Item 3' }
   ];
-  targetItems = [];
-  targetItemsA = [];
-  targetItemsB = [];
+  targetItems: any[] = [];
+  targetItemsA: any[] = [];
+  targetItemsB: any[] = [];
 
   sourceNestedItems = [
-    { label: 'Item 1, no children', children: [] },
+    { label: 'Item 1, no children', children: [] as any[] },
     { label: 'Item 2', children: [
         { label: 'no' },
         { label: 'children' }
@@ -53,25 +53,29 @@ export class AppComponent {
     },
     { label: 'Item 3, can\'t have children' }
   ];
-  targetNestedItems = [];
+  targetNestedItems: any[] = [];
 
   sourceBuilderTools = [
     { name: 'Section', children: [], inputType: 'section', icon: 'section', class: 'wide' },
     { name: 'A String', inputType: 'string', icon: 'field-text', class: 'half' },
     { name: 'A Number', inputType: 'number', icon: 'field-numeric', class: 'half' }
   ];
-  targetBuilderTools = [];
+  targetBuilderTools: any[] = [];
 
-  droppableItemClass = item => `${item.class} ${item.inputType}`;
+  droppableItemClass = (item: any) => `${item.class} ${item.inputType}`;
 
-  builderDrag(e) {
+  builderDrag(e: any) {
     const item = e.value;
     item.data = item.inputType === 'number' ?
       (Math.random() * 100) | 0 :
       Math.random().toString(36).substring(20);
   }
 
-  log(e) {
+  log(e: any) {
     console.log(e.type, e);
+  }
+
+  canMove(e: any): boolean {
+    return e.indexOf('Disabled') === -1;
   }
 }