+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Support strings as droppableItemClass
 * Improved component reactivity
+* Feature: add moves input to ngxDraggable and ContainerComponent
 
 --------------------